backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
chardet==5.2.0
click==8.1.7
colorama==0.4.6
coverage==7.6.8
dnspython==2.2.1
email_validator==2.2.0
eval_type_backport==0.2.0
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jsonpath-python==1.0.6
lxml==5.3.0
mistralai==1.2.3
mongomock==4.3.0
mypy-extensions==1.0.0
packaging==24.2
phonenumbers==8.13.48
pillow==11.0.0
pluggy==1.5.0
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.10.1
PyPDF2==3.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
reportlab==4.2.5
sentinels==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing-inspect==0.9.0
typing_extensions==4.12.2
unicorn==2.1.1
uvicorn==0.32.0
wcwidth==0.2.13
